MVNMDVAL-8: Add owner exception for mod-pubsub circular dependency
Allow mod-pubsub to define the mod-audit and mod-remote-storage API endpoint permissions.
This solves the circular dependency when assigning permissions to its system user
when enabling the module for a tenant because mod-audit and mod-remote-storage are not enabled at that time.

We allow this exception because https://github.com/folio-org/mod-pubsub is deprecated and will be removed soon.